+++ TaskGame/src/main/java/net/fred/taskgame/widget/WidgetConfigurationActivity.java
@@ -78,7 +78,7 @@ public class WidgetConfigurationActivity extends Activity {
             @Override
             public void onClick(View v) {
 
-                int categoryId = -1;
+                long categoryId = -1;
                 if (mRadioGroup.getCheckedRadioButtonId() != R.id.widget_config_tasks) {
                     Category tag = (Category) categorySpinner.getSelectedItem();
                     categoryId = tag.id;
